requirements.txt
fastapi==0.115.12
httpx==0.28.1
isaacus==0.21.0
python-dotenv==1.1.0
uvicorn==0.34.2
